cleanup, clarity
