+++ search/process/tests.py
@@ -38,9 +38,9 @@ class TestTransformMetdata(TestCase):
             ]
         })
         doc = transform.to_search_document(meta)
-        self.assertEqual(doc.authors[0]['first_name'], 'B Ivan')
+        self.assertEqual(doc.authors[0]['first_name'], 'B. Ivan')
         self.assertEqual(doc.authors[0]['last_name'], 'Dole')
-        self.assertEqual(doc.authors[0]['full_name'], 'B Ivan Dole',
+        self.assertEqual(doc.authors[0]['full_name'], 'B. Ivan Dole',
                          "full_name should be generated from first_name and"
                          " last_name")
         self.assertEqual(doc.authors[0]['initials'], ["B", "I"],
@@ -67,9 +67,9 @@ class TestTransformMetdata(TestCase):
             ]
         })
         doc = transform.to_search_document(meta)
-        self.assertEqual(doc.owners[0]['first_name'], 'B Ivan')
+        self.assertEqual(doc.owners[0]['first_name'], 'B. Ivan')
         self.assertEqual(doc.owners[0]['last_name'], 'Dole')
-        self.assertEqual(doc.owners[0]['full_name'], 'B Ivan Dole',
+        self.assertEqual(doc.owners[0]['full_name'], 'B. Ivan Dole',
                          "full_name should be generated from first_name and"
                          " last_name")
         self.assertEqual(doc.owners[0]['initials'], ["B", "I"],
@@ -216,22 +216,22 @@ class TestTransformMetdata(TestCase):
                          meta.secondary_classification)
 
     def test_title(self):
-        """Field ``title`` is populated from ``title``."""
+        """Field ``title`` is populated from ``title_utf8``."""
         meta = DocMeta(**{
             'paper_id': '1234.56789',
-            'title': 'foo title'
+            'title_utf8': 'foo title'
         })
         doc = transform.to_search_document(meta)
         self.assertEqual(doc.title, 'foo title')
 
     def test_title_utf8(self):
-        """Field ``title_utf8`` is populated from ``title_utf8``."""
+        """Field ``title`` is populated from ``title_utf8``."""
         meta = DocMeta(**{
             'paper_id': '1234.56789',
             'title_utf8': 'foö title'
         })
         doc = transform.to_search_document(meta)
-        self.assertEqual(doc.title_utf8, 'foö title')
+        self.assertEqual(doc.title, 'foö title')
 
     def test_source(self):
         """Field ``source`` is populated from ``source``."""
