+++ search/services/index/util.py
@@ -6,7 +6,9 @@ from string import punctuation
 
 from elasticsearch_dsl import Search, Q, SF
 
+from search.domain import Query
 from .exceptions import QueryError
+
 
 # We'll compile this ahead of time, since it gets called quite a lot.
 STRING_LITERAL = re.compile(r"(['\"][^'\"]*['\"])")
@@ -18,11 +20,9 @@ TEXISM = re.compile(r'(\$[^\$]+\$)')
 MAX_RESULTS = 10_000
 """This is the maximum result offset for pagination."""
 
-HIGHLIGHT_TAG_OPEN = '<span class="search-hit mathjax">'
-HIGHLIGHT_TAG_CLOSE = '</span>'
-
 SPECIAL_CHARACTERS = ['+', '-', '=', '&&', '||', '>', '<', '!', '(', ')', '{',
                       '}', '[', ']', '^', '~', ':', '\\', '/']
+DEFAULT_SORT = ['_score', '-announced_date_first', '_doc']
 
 
 def wildcardEscape(querystring: str) -> Tuple[str, bool]:
@@ -96,6 +96,7 @@ def escape(term: str) -> str:
 
 
 def strip_punctuation(s: str) -> str:
+    """Remove all punctuation characters from a string."""
     return ''.join([c for c in s if c not in punctuation])
 
 
@@ -103,3 +104,15 @@ def remove_single_characters(term: str) -> str:
     """Remove any single characters in the search string."""
     return ' '.join([part for part in term.split()
                      if len(strip_punctuation(part)) > 1])
+
+
+def sort(query: Query, search: Search) -> Search:
+    """Apply sorting to a :class:`.Search`."""
+    if not query.order:
+        sort_params = DEFAULT_SORT
+    else:
+        direction = '-' if query.order.startswith('-') else ''
+        sort_params = [query.order, f'{direction}paper_id_v']
+    if sort_params is not None:
+        search = search.sort(*sort_params)
+    return search
